将前端文件置入静态文件夹，更改前后端交互接口以同时适配在线与本地
