+++ src/main/java/com/codex/taxitrajectory/config/WebConfig.java
@@ -11,7 +11,7 @@ public class WebConfig implements WebMvcConfigurer {
     @Override
     public void addCorsMappings(CorsRegistry registry) {
         registry.addMapping("/**") // 匹配所有的URL路径
-                .allowedOrigins("http://localhost:63342") // 允许的来源
+                .allowedOrigins("http://fenrirchen.com", "http://localhost:63342") // 允许的来源
                 .allowedMethods("GET", "POST", "PUT", "DELETE", "OPTIONS") // 允许的HTTP方法
                 .allowedHeaders("*") // 允许所有头信息
                 .allowCredentials(true); // 是否允许发送Cookie
