+++ src/main/resources/static/js/features/f2-zoom.js
