+++ src/main/resources/static/js/utils/map-utils.js
