+++ src/main/resources/static/taxi-trajectory.html
